Software/AIris-Final-App-Old/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
opencv-python-headless==4.10.0.84
ultralytics==8.3.0
torch>=2.0.0
torchvision>=0.15.0
mediapipe>=0.10.11,<0.11.0
Pillow==10.4.0
groq==0.11.0
python-dotenv==1.0.1
transformers==4.46.0
torchaudio>=2.0.0
gTTS==2.5.1
pyyaml==6.0.2
numpy>=1.23.0,<2.0.0
pydub>=0.25.1

